LayoutTests/platform/wk2/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2-expected.txt

Button1  Button2  Invoker0  Invoker1  Button3  Button4
Invoker  after
Show popover
Toggle popover Other focusable element

FAIL Popover focus navigation assert_true: popover1 should be invoked by invoker1 expected true got false
FAIL Circular reference tab navigation assert_equals: circular reference: Step 1 (backwards) expected Element node <button id="circular0" popovertarget="popover4" tabindex=... but got Element node <button id="button4" tabindex="0">Button4</button>
PASS Popover focus returns when popover is hidden by invoker
PASS Popover focus only returns to invoker when focus is within the popover
